paylocity-automation-framework/requirements.txt
# ===============================
# Test Runner & Assertions
# ===============================
pytest==8.0.0
pytest-xdist==3.5.0

# ===============================
# UI Automation
# ===============================
selenium==4.18.1
webdriver-manager==4.0.1

# ===============================
# API Automation
# ===============================
requests==2.31.0

# ===============================
# Reporting
# ===============================
allure-pytest==2.13.2

# ===============================
# Configuration Management
# ===============================
python-dotenv==1.0.1
PyYAML==6.0.1

# ===============================
# Utilities
# ===============================
loguru==0.7.2